+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## 2.0.0 - 2023-10-25
+## 3.0.0 - 2024-03-10
+
+### Changed
+
+- Requires `innmind/operating-system:~5.0`
+- Requires `innmind/io:~2.7`
+- Requires `innmind/http-parser:~2.1`
+- `Innmind\Async\HttpServer\Server::of()` now accept either a `Innmind\IO\Sockets\Server` or `Innmind\IO\Sockets\Server\Pool`
+
+## 2.0.0 - 2023-11-25
 
 ### Changed